リクエストパラーメータでidを直接渡している箇所が、自分が閲覧可能なものかチェック処理をいれる
closed #156

+++ app/controllers/api/v1/user/bank_accounts_controller.rb
@@ -44,6 +44,10 @@ class Api::V1::User::BankAccountsController < ApplicationController
 
   def update
     account_id = params[:id].to_i
+    if disallowed_at_bank_ids?([account_id])
+      render_disallowed_financier_ids && return
+    end
+
     if @current_user.try(:at_user).try(:at_user_bank_accounts).pluck(:id).include?(account_id)
       require_group && return if params[:share] == true
       account = Entities::AtUserBankAccount.find account_id
@@ -64,6 +68,10 @@ class Api::V1::User::BankAccountsController < ApplicationController
 
   def destroy
     account_id = params[:id].to_i
+    if disallowed_at_bank_ids?([account_id])
+      render_disallowed_financier_ids && return
+    end
+
     if @current_user.try(:at_user).try(:at_user_bank_accounts).pluck(:id).include?(account_id)
       Services::AtUserService.new(@current_user).delete_account(Entities::AtUserBankAccount, [account_id])
     end
